% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APACHE2.R
\name{APACHEII_missingness}
\alias{APACHEII_missingness}
\title{Summarize APACHEII missingness}
\usage{
APACHEII_missingness(APACHEII_scores)
}
\arguments{
\item{APACHEII_scores}{is a dataframe with APACHEII scores}
}
\value{
dataframe with APACHEII score & component missingness
}
\description{
The APACHEII_missingness function summarizes missingness of
APACHEII points & scores that might contextualize estimates.
}
