% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APACHE2.R
\name{APACHEII_mortality}
\alias{APACHEII_mortality}
\title{Estimate in-hospital mortality}
\usage{
APACHEII_mortality(APACHEII_scores)
}
\arguments{
\item{APACHEII_score}{is a dataframe with APACHEII scores}

\item{op_status}{is a column name that indicates operative status,
which should be coded as 0 = non-operative, 1 = post-operative.}
}
\value{
dataframe with APACHE II score from APACHEII_points function.
}
\description{
This function takes a APACHE II scores calculated with the
APACHEII_points function and optional parameter to indicate
operative status and estimates in-hospital mortality rates.
}
\details{
The output is a dataframe with APACHE II scores, non-operative,
and operative mortality unless operative status is specified.
}
