% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APACHE2.R
\name{APACHEII_points}
\alias{APACHEII_points}
\title{Assign APACHE II points & calculate score}
\usage{
APACHEII_points(sepsis_data)
}
\arguments{
\item{sepsis_data}{is a dataframe with critical care data}
}
\value{
dataframe with APACHE II score & component point values.
}
\description{
This function takes a cleaned dataframe with critical care data
and assumes variable names as defined in the COVID project data
dictionary but can be modified with some elbow grease.
}
\details{
The APACHEII_points function classifies data, assigns an APACHEII
score, and returns a data frame with these results.
}
