% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOFAs.R
\name{SOFA_missingness}
\alias{SOFA_missingness}
\title{Summarize SOFA & qSOFA missingness}
\usage{
SOFA_missingness(covid_points)
}
\arguments{
\item{sepsis_data}{is a dataframe with critical care data}
}
\value{
dataframe with SOFA & qSOFA score & component missingness
}
\description{
The SOFA_missingness function summarizes missingness of SOFA &
qSOFA parameters that might contextualize variation in estimates.
}
