% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOFAs.R
\name{SOFA_mortality}
\alias{SOFA_mortality}
\title{Predict mortality risk with SOFA & qSOFA scores}
\usage{
SOFA_mortality(SOFA_scores)
}
\arguments{
\item{SOFA_scores}{is a dataframe with scores from SOFA_points}
}
\value{
mortality dataframe with SOFA & qSOFA scores and mortality risk
}
\description{
The SOFA_mortality function predicts mortality based on
SOFA and qSOFA scores calculated with the SOFA_points function.
}
