% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOFAs.R
\name{SOFA_points}
\alias{SOFA_points}
\title{Calculate SOFA & qSOFA scores}
\usage{
SOFA_points(sepsis_data)
}
\arguments{
\item{sepsis_data}{is a dataframe with critical care data}
}
\value{
dataframe with SOFA & qSOFA scores & components
}
\description{
This function takes a cleaned dataframe with critical care data
and assumes variable names as defined in the COVID project data
dictionary but can be modified with some elbow grease.
}
\details{
The SOFA_points function classifies data, assigns a SOFA
and qSOFA score, and returns a data frame with these results.
}
